requirements.txt

aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.13.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
redis-dict==2.5.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
XlsxWriter==3.2.0
yarl==1.9.4